requirements.txt
requests
beautifulsoup4
schedule
pyyaml
